backend/requirements.txt
flask
flask-cors
firebase-admin
deepface
opencv-python
python-dotenv
numpy
tf-keras
